requirements.txt
google-generativeai
tweepy
python-dotenv
requests
schedule